2.17: La guía telefónica, paso 9
https://fullstackopen.com/es/part2/alterando_datos_en_el_servidor#ejercicios-2-15-2-18

+++ part2/phonebook/src/App.js
@@ -51,9 +51,10 @@ const App = () => {
   }
 
   const handlePersonDeleted = id => {
-    personsService.remove(id).then(() => {
-      setPersons(persons.filter(person => person.id !== id))
-    })
+    if (window.confirm(`Delete ${persons.find(p => p.id === id).name}?`))
+      personsService.remove(id).then(() => {
+        setPersons(persons.filter(person => person.id !== id))
+      })
   }
 
   return (
